requirements.txt
# ===============================
# Python
# ===============================
# python==3.10

# ===============================
# Deep Learning
# ===============================
torch>=2.0
torchvision

# ===============================
# Detection
# ===============================
ultralytics>=8.1,<9.0

# ===============================
# Recognition
# ===============================
timm==0.9.12
transformers==4.36.2
easyocr==1.7.1

# ===============================
# Utils
# ===============================
opencv-python
numpy
pandas
matplotlib
scikit-learn
python-Levenshtein
Pillow
PyYAML
